Restore missing dbt-sqlserver specific code to test harness (#55)
* Missing data column in feature file

* Add dbt-sqlserver adapter to Pipfile and Pipfile.lock

* First attempt at restoring sqlserver specific code into test harness

Tested with test/features/hubs/sqlserver/sqlserver_hubs.feature, some tests currently failing and I am investigating.

* Fix for failing pit macro call

Issue reported by Flynn 01/09/2021.

* All missing sqlserver specific code now restored

The test harness is now running tests in sqlserver specific feature files.
In shared_steps.py all sqleserver code is executed conditionally.

* sqlserver hub tests now passing

* PIT macro new TODO

Should the length of the ghost_pk zero hash be determined by the hashing option being used, i.e. MD5 = 16, SHA = 32 ?

* Reformat code

* Revision per PR comment

+++ test/dbtvault_harness_utils.py
@@ -27,9 +27,9 @@ def platform():
 
         with open(test.INVOKE_YML_FILE) as config:
             config_dict = yaml.safe_load(config)
-            plt = config_dict.get('platform')
+            plt = config_dict.get('platform').lower()
 
-            if plt.lower() not in test.AVAILABLE_PLATFORMS:
+            if plt not in test.AVAILABLE_PLATFORMS:
                 test.logger.error(f"Platform must be set to one of: {', '.join(test.AVAILABLE_PLATFORMS)} "
                                   f"in '{test.INVOKE_YML_FILE}'")
                 sys.exit(0)
@@ -114,15 +114,19 @@ def clean_target():
     shutil.rmtree(test.TEST_PROJECT_ROOT / 'target', ignore_errors=True)
 
 
-def clean_csv():
+def clean_csv(model_name=None):
     """
     Deletes csv files in csv folder.
     """
 
-    delete_files = [file for file in glob.glob(str(test.CSV_DIR / '*.csv'), recursive=True)]
+    if model_name:
+        delete_files = [test.CSV_DIR / f"{model_name.lower()}.csv"]
+    else:
+        delete_files = [file for file in glob.glob(str(test.CSV_DIR / '*.csv'), recursive=True)]
 
     for file in delete_files:
-        os.remove(file)
+        if os.path.isfile(file):
+            os.remove(file)
 
 
 def clean_models():
@@ -133,7 +137,8 @@ def clean_models():
     delete_files = [file for file in glob.glob(str(test.TEST_MODELS_ROOT / '*.sql'), recursive=True)]
 
     for file in delete_files:
-        os.remove(file)
+        if os.path.isfile(file):
+            os.remove(file)
 
 
 def create_dummy_model():
